video_processor/requirements.txt
yt-dlp==2024.1.1
ffmpeg-python==0.2.1
openai==1.3.0
python-dotenv==1.0.0
pytest==7.4.3
hypothesis==6.92.1
